VMPARKING module uses SQLITE3
After test, pubsub as big database is not scalable.
In order to improve perfs of mass parking/unparking
the module is now using a sqlite3 file.
This file need to be shared (by default it will be set into /vm/shared_parking.sqlite3)

You need to add [VMPARKING] conf token database to set the path of the databse if you
use a previous version